+++ client/src/components/Product/Product.module.css
@@ -13,11 +13,11 @@
 .productContainer {
   width: 100%;
 }
-.productContainer > img {
-  /* height: 70%;
+/* .productContainer > img {
+  height: 70%;
   max-width: 100%;
-  border: solid 1px black; */
-}
+  border: solid 1px black; 
+} */
 .image_container {
   border-radius: 6px;
   width: 100%;
@@ -38,13 +38,13 @@
   cursor: pointer;
 }
 
-.productContainer > h3 {
-  /* width: 100%;
+/* .productContainer > h3 {
+  width: 100%;
   font-weight: 500;
   height: 8%;
   margin: 0;
-  padding-left: 7px; */
-}
+  padding-left: 7px;
+} */
 .title,
 .p {
   color: #333;
@@ -63,16 +63,16 @@
   font-size: 14px;
   margin-bottom: 16px;
 }
-.productContainer > label {
-  /* font-size: 13px;
+/* .productContainer > label {
+  font-size: 13px;
   height: 10%;
-  padding-left: 7px; */
-}
+  padding-left: 7px; 
+} */
 
 .description {
   padding: 16px;
 }
-.btnAddCart {
+.btnAddCart > button {
   /* width: 100%;
   height: 12%;
   padding: 0 5px 0 5px; */
@@ -92,9 +92,6 @@
   right: 16px;
   bottom: 16px;
 }
-.btnAddCart > button {
-  /* width: 100%; */
-}
-.btnAddCart:hover {
+.btnAddCart > button:hover {
   cursor: pointer;
 }
